requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
phonenumbers==8.13.35
pillow==10.3.0
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0

cryptography~=42.0.2
typing_extensions~=4.10.0
cffi~=1.16.0
defusedxml~=0.8.0rc2
pip~=24.0
automium_web~=0.1.1
Jinja2~=3.1.3
colorama~=0.4.6
PyYAML~=6.0.1
numpy~=1.26.4
setuptools~=65.5.0
packaging~=24.0
django-ckeditor~=6.7.1
django-crispy-forms~=1.14.0
uritemplate~=4.1.1
inflection~=0.5.1
pytz~=2024.1
django-modeltranslation~=0.18.11
legacy~=0.1.7